Add a file that contains the class Pricing
Immplemented a class Princing that accepts two arguments (Number) and curency (Currency). The getter and setter methods of the attributes where also implemented. diplayFullPrice which returns the attributes in the following format  was also implemented along side the static convertPrice method that accepts two arguments amount and conversionrate that is used for converting an amount based on the Conversion rate