Add a file that contains a function  and a const instance of WeakMap
Implemented a function queryAPI that accepts an endpoint argument: endpoint(JSON). The function has fature of counting how may times it is being called and when it subpasses or is equal to 5 and error is thrown. The weakMap object instance helps in storing the count how many times the function is being called